pyproject.toml
[project]
name = "analyst-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "openai-chatkit>=0.0.2",
    "ruff>=0.13.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
